+++ packages/versioning/src/extensions/reentry-status/index.ts
@@ -12,5 +12,6 @@ export * from './roadmap-renderer';
 export * from './git-context';
 export * from './reentry-status-manager';
 export * from './status-renderer';
+export * from './validation';
 import extension from './extension';
 export default extension;